Create Shashank has punished - count number of vowels
Shashank has been punished by his teacher for not doing his homework. He has been given a set of words and he has to find the no. of vowels in each of these words. Shashank knows that doing this manually will be a tedious task and asks you to help him.
Write a function that accepts one word(without space) and counts the no. of

vowels(a,e,i,o,u) in that word.

Example 1:
If you pass "Wipro", it should return the value 2 as there are two vowels in this
word(i and o).

Example 2:
If you pass "tEchnologies", it should return the value 5, as there are 5 vowels in this word (E,0,0,i,e).